Switched Stats from Int to Double and edit client to changes of back end

+++ Tichu/Views/Share/ShareStatsView.swift
@@ -13,7 +13,7 @@ struct ShareStatsView: View{
     var userName: String
     var userImageData: Data?
     var elo: Double?
-    var winnerPercentage: Int
+    var winnerPercentage: Double
     var tichuMaster: Double
     let accentCo: Color
